Scoring zone config in configure assemblies panel
