+++ fission/src/ui/panels/configuring/assembly-config/interfaces/ConfigureGamepiecePickupInterface.tsx
@@ -2,25 +2,26 @@ import * as THREE from "three"
 import { useCallback, useEffect, useMemo, useState } from "react"
 import SelectButton from "@/components/SelectButton"
 import TransformGizmos from "@/ui/components/TransformGizmos"
+import World from "@/systems/World"
 import Slider from "@/ui/components/Slider"
 import Jolt from "@barclah/jolt-physics"
 import PreferencesSystem from "@/systems/preferences/PreferencesSystem"
 import MirabufSceneObject, { RigidNodeAssociate } from "@/mirabuf/MirabufSceneObject"
-import World from "@/systems/World"
+import { RigidNodeId } from "@/mirabuf/MirabufParser"
 import {
     Array_ThreeMatrix4,
     JoltMat44_ThreeMatrix4,
     ReactRgbaColor_ThreeColor,
     ThreeMatrix4_Array,
 } from "@/util/TypeConversions"
-import { useTheme } from "@/ui/ThemeContext"
 import LabeledButton, { LabelPlacement } from "@/ui/components/LabeledButton"
-import { RigidNodeId } from "@/mirabuf/MirabufParser"
-import { ConfigurationSavedEvent } from "./ConfigureAssembliesPanel"
+import { useTheme } from "@/ui/ThemeContext"
+import React from "react"
+import { ConfigurationSavedEvent } from "../ConfigureAssembliesPanel"
 
 // slider constants
-const MIN_VELOCITY = 0.0
-const MAX_VELOCITY = 20.0
+const MIN_ZONE_SIZE = 0.1
+const MAX_ZONE_SIZE = 1.0
 
 /**
  * Saves ejector configuration to selected robot.
@@ -45,13 +46,8 @@ const MAX_VELOCITY = 20.0
  * @param selectedRobot Selected robot to save data to.
  * @param selectedNode Selected node that configuration is relative to.
  */
-function save(
-    ejectorVelocity: number,
-    gizmo: TransformGizmos,
-    selectedRobot: MirabufSceneObject,
-    selectedNode?: RigidNodeId
-) {
-    if (!selectedRobot?.ejectorPreferences || !gizmo) {
+function save(zoneSize: number, gizmo: TransformGizmos, selectedRobot: MirabufSceneObject, selectedNode?: RigidNodeId) {
+    if (!selectedRobot?.intakePreferences || !gizmo) {
         return
     }
 
@@ -62,39 +58,41 @@ function save(
         return
     }
 
-    const gizmoTransformation = gizmo.mesh.matrixWorld
+    const translation = new THREE.Vector3(0, 0, 0)
+    const rotation = new THREE.Quaternion(0, 0, 0, 1)
+    gizmo.mesh.matrixWorld.decompose(translation, rotation, new THREE.Vector3(1, 1, 1))
+
+    const gizmoTransformation = new THREE.Matrix4().compose(translation, rotation, new THREE.Vector3(1, 1, 1))
     const robotTransformation = JoltMat44_ThreeMatrix4(World.PhysicsSystem.GetBody(nodeBodyId).GetWorldTransform())
     const deltaTransformation = gizmoTransformation.premultiply(robotTransformation.invert())
 
-    selectedRobot.ejectorPreferences.deltaTransformation = ThreeMatrix4_Array(deltaTransformation)
-    selectedRobot.ejectorPreferences.parentNode = selectedNode
-    selectedRobot.ejectorPreferences.ejectorVelocity = ejectorVelocity
+    selectedRobot.intakePreferences.deltaTransformation = ThreeMatrix4_Array(deltaTransformation)
+    selectedRobot.intakePreferences.parentNode = selectedNode
+    selectedRobot.intakePreferences.zoneDiameter = zoneSize
 
     PreferencesSystem.savePreferences()
 }
 
-interface ConfigEjectorProps {
+interface ConfigPickupProps {
     selectedRobot: MirabufSceneObject
 }
 
-const ConfigureShotTrajectoryInterface: React.FC<ConfigEjectorProps> = ({ selectedRobot }) => {
+const ConfigureGamepiecePickupInterface: React.FC<ConfigPickupProps> = ({ selectedRobot }) => {
     const { currentTheme, themes } = useTheme()
     const theme = useMemo(() => {
         return themes[currentTheme]
     }, [currentTheme, themes])
 
     const [selectedNode, setSelectedNode] = useState<RigidNodeId | undefined>(undefined)
-    const [ejectorVelocity, setEjectorVelocity] = useState<number>((MIN_VELOCITY + MAX_VELOCITY) / 2.0)
+    const [zoneSize, setZoneSize] = useState<number>((MIN_ZONE_SIZE + MAX_ZONE_SIZE) / 2.0)
     const [transformGizmo, setTransformGizmo] = useState<TransformGizmos | undefined>(undefined)
 
     const saveEvent = useCallback(() => {
         if (transformGizmo && selectedRobot) {
-            save(ejectorVelocity, transformGizmo, selectedRobot, selectedNode)
-            const currentGp = selectedRobot.activeEjectable
-            selectedRobot.SetEjectable(undefined, true)
-            selectedRobot.SetEjectable(currentGp)
+            save(zoneSize, transformGizmo, selectedRobot, selectedNode)
+            selectedRobot.UpdateIntakeSensor()
         }
-    }, [transformGizmo, selectedRobot, selectedNode, ejectorVelocity])
+    }, [transformGizmo, selectedRobot, selectedNode, zoneSize])
 
     useEffect(() => {
         ConfigurationSavedEvent.Listen(saveEvent)
@@ -104,16 +102,24 @@ const ConfigureShotTrajectoryInterface: React.FC<ConfigEjectorProps> = ({ select
         }
     }, [saveEvent])
 
+    useEffect(() => {
+        if (!transformGizmo) {
+            return
+        }
+
+        transformGizmo.mesh.scale.set(zoneSize, zoneSize, zoneSize)
+    }, [zoneSize, transformGizmo])
+
     // Not sure I like this, but made it a state and effect rather than a memo to add the cleanup to the end
     useEffect(() => {
-        if (!selectedRobot?.ejectorPreferences) {
+        if (!selectedRobot?.intakePreferences) {
             setTransformGizmo(undefined)
             return
         }
 
         const gizmo = new TransformGizmos(
             new THREE.Mesh(
-                new THREE.ConeGeometry(0.1, 0.4, 4).rotateX(Math.PI / 2.0).translate(0, 0, 0.2),
+                new THREE.SphereGeometry(0.5),
                 World.SceneRenderer.CreateToonMaterial(ReactRgbaColor_ThreeColor(theme.HighlightSelect.color))
             )
         )
@@ -121,12 +127,11 @@ const ConfigureShotTrajectoryInterface: React.FC<ConfigEjectorProps> = ({ select
         ;(gizmo.mesh.material as THREE.Material).depthTest = false
         gizmo.AddMeshToScene()
         gizmo.CreateGizmo("translate", 1.5)
-        gizmo.CreateGizmo("rotate", 2.0)
 
-        const deltaTransformation = Array_ThreeMatrix4(selectedRobot.ejectorPreferences.deltaTransformation)
+        const deltaTransformation = Array_ThreeMatrix4(selectedRobot.intakePreferences.deltaTransformation)
 
         let nodeBodyId = selectedRobot.mechanism.nodeToBody.get(
-            selectedRobot.ejectorPreferences.parentNode ?? selectedRobot.rootNodeId
+            selectedRobot.intakePreferences.parentNode ?? selectedRobot.rootNodeId
         )
         if (!nodeBodyId) {
             // In the event that something about the id generation for the rigid nodes changes and parent node id is no longer in use
@@ -143,15 +148,16 @@ const ConfigureShotTrajectoryInterface: React.FC<ConfigEjectorProps> = ({ select
         setTransformGizmo(gizmo)
 
         return () => {
+            console.log("remove gizmos")
             gizmo.RemoveGizmos()
             setTransformGizmo(undefined)
         }
     }, [selectedRobot, theme])
 
     useEffect(() => {
-        if (selectedRobot?.ejectorPreferences) {
-            setEjectorVelocity(selectedRobot.ejectorPreferences.ejectorVelocity)
-            setSelectedNode(selectedRobot.ejectorPreferences.parentNode)
+        if (selectedRobot?.intakePreferences) {
+            setZoneSize(selectedRobot.intakePreferences.zoneDiameter)
+            setSelectedNode(selectedRobot.intakePreferences.parentNode)
         } else {
             setSelectedNode(undefined)
         }
@@ -193,13 +199,13 @@ const ConfigureShotTrajectoryInterface: React.FC<ConfigEjectorProps> = ({ select
 
             {/* Slider for user to set velocity of ejector configuration */}
             <Slider
-                min={MIN_VELOCITY}
-                max={MAX_VELOCITY}
-                value={ejectorVelocity}
-                label="Velocity"
+                min={MIN_ZONE_SIZE}
+                max={MAX_ZONE_SIZE}
+                value={zoneSize}
+                label="Size"
                 format={{ minimumFractionDigits: 2, maximumFractionDigits: 2 }}
                 onChange={(_, vel: number | number[]) => {
-                    setEjectorVelocity(vel as number)
+                    setZoneSize(vel as number)
                 }}
                 step={0.01}
             />
@@ -216,7 +222,7 @@ const ConfigureShotTrajectoryInterface: React.FC<ConfigEjectorProps> = ({ select
                         transformGizmo.mesh.position.setFromMatrixPosition(robotTransformation)
                         transformGizmo.mesh.rotation.setFromRotationMatrix(robotTransformation)
                     }
-                    setEjectorVelocity((MIN_VELOCITY + MAX_VELOCITY) / 2.0)
+                    setZoneSize((MIN_ZONE_SIZE + MAX_ZONE_SIZE) / 2.0)
                     setSelectedNode(selectedRobot?.rootNodeId)
                 }}
             />
@@ -224,4 +230,4 @@ const ConfigureShotTrajectoryInterface: React.FC<ConfigEjectorProps> = ({ select
     )
 }
 
-export default ConfigureShotTrajectoryInterface
+export default ConfigureGamepiecePickupInterface
